+++ src/Cirreum.Secrets.Azure/Configuration/AzureKeyVaultInstanceSettings.cs
@@ -18,7 +18,18 @@ public class AzureKeyVaultInstanceSettings
 	/// If you have a secret <see cref="Uri"/>, use <see cref="KeyVaultSecretIdentifier"/> to parse the <see cref="KeyVaultSecretIdentifier.VaultUri"/> and other information.
 	/// You should validate that this URI references a valid Key Vault resource. See <see href="https://aka.ms/azsdk/blog/vault-uri"/> for details.
 	/// </summary>
-	public Uri? VaultUri { get; set; }
+	protected internal Uri? VaultUri { get; set; }
+
+	/// <summary>
+	/// Determines which credential is used to authenticate to the vault.
+	/// </summary>
+	public CredentialMode CredentialMode { get; set; } = CredentialMode.Default;
+
+	/// <summary>
+	/// The client ID of a user-assigned managed identity. Only used when <see cref="CredentialMode"/> is
+	/// <see cref="CredentialMode.ManagedIdentity"/>. If not supplied, the system-assigned managed identity is used.
+	/// </summary>
+	public string? ManagedIdentityClientId { get; set; }
 
 	public override void ParseEndPoint() {
 		if (!string.IsNullOrEmpty(this.Endpoint) &&